WIP : attempt to support for Volta with Pascal thread scheduling turned on. see GENCODE_SM70 arch mismatch and read https://devblogs.nvidia.com/using-cuda-warp-l5Devel-primitives
instrinsics like __shfl() __shfl_*() and __ballot() are deprecated in favor of __shfl_sync( mask, ...) __shfl_*_sync( mask, ...) and __ballot_sync( mask, ... )
this is not as simple as always using the full mask ... as thread have to be launched to match the complete mask for a warp not to HANG !!

idem cudaThreadSynchronize() is deprecated ...

for now, using Pascal thread scheduling mode by passing arch=compute_60 for code=sm_70 ...
Hoping the results are correct.

A real port of CUDPP to CUDA 9.x cooperative threading model may very well be required.